feat: warden role list + skills catalog uses typed introspection
Renames the agent-facing role-introspection CLI from "warden roles"
(top-level) to "warden role list" (subcommand of a new "role" parent
group), mirroring "warden provider list". Safe to rename in place: the
original surface lives only on [Unreleased] and never shipped.

Updates the skill catalog to use the typed surfaces everywhere - the
generic "warden list sys/providers" becomes "warden provider list",
and "warden roles" becomes "warden role list". Fixes the discovery
SKILL's example response shape to match the actual array-of-records
output of "warden provider list -o json".

E2E tests adjusted: TestRolesCLI_JWTPromotesToBearer and
TestRolesCLI_AuthPathFilter now invoke "role list" instead of "roles".
