+++ CHANGELOG.md
@@ -18,7 +18,7 @@ All notable changes to Warden are documented in this file.
 
 - **`--json` payload on every mutating typed command.** Adds a `--json` / `-j` flag accepting a literal JSON object (`--json '{"type":"aws",...}'`), a file reference (`--json @aws.json`), or `-` for stdin (`cat aws.json | warden ... --json -`) on every typed mutating command: `cred source create/update`, `cred spec create/update`, `auth enable`, `audit enable`, `namespace create/update`, `provider enable`. Mutually exclusive with the per-typed-flag surface (`--type`, `--config`, etc.); combining the two errors with exit `2` (`usage`). For commands where the mount path can be derived from the type (`auth enable`, `audit enable`, `provider enable`), the path is taken from the positional argument when given, otherwise from the payload's `type` field. Composes cleanly with `--dry-run`: agents can iterate `build payload → --dry-run → fix typos → real call` against any mutating endpoint without writing the typed-flag flattening glue. The previously-required `--type`, `--source`, `--rotation-period`, `--metadata` flags become required-unless-`--json`. Server-side fix bundled: `handleCredentialSourceUpdate` now preserves `RotationPeriod` from the existing source on update — without this, every `hvault` source update failed validation (`rotation_period is required`), and non-`hvault` updates silently stripped their rotation configuration.
 - **Global `--dry-run` / `-D` flag — local schema validation, no server round-trip.** The flag never sends a request: it fetches the operation's schema from `/v1/sys/schema?path=<path>` (cached per process), validates the agent's payload structurally — required fields, unknown field names with Levenshtein "did you mean" hints, JSON type matches, enum/allowed-value checks — and either exits `0` with a `{"path", "validated": true, "dry_run": true}` envelope or exits `3` (`invalid_input`) with the structured error envelope listing every problem at once. Honors `WARDEN_DRY_RUN` (any value other than empty / `0` / `false` / `no` / `off`). Wired into every mutating CLI surface: `warden write` / `warden delete`, `warden auth enable/disable`, `warden audit enable/disable`, `warden provider enable/disable`, `warden cred source create/update/delete`, `warden cred spec create/update/delete`, `warden namespace create/update/delete`, `warden policy write/delete`. The validator catches the failure mode that matters most for agents: hallucinated parameters. Authz/RBAC and conflict checks are NOT verified because the request never reaches the server — those failures surface only when the agent runs the real command, and the cost of that is "you get an error response" rather than data loss. Server-side enhancement bundled: `projectOASToPath` now matches concrete paths against templated keys (`sys/cred/sources/my-aws` finds `sys/cred/sources/{name}`), so `warden schema PATH` works without agents knowing placeholder names.
-- **`warden roles` CLI command** — agent-facing role introspection over the server's `/v1/sys/introspect/roles` aggregator. Lists the `{name, description, auth_path}` of every role the caller's identity (JWT bearer or TLS client cert) can assume across the namespace's auth mounts. Composes with `--output` and `--fields`; per-mount failures surface on stderr without changing the exit code so a partial-failure mount can't mask the rest of the namespace's roles. `--auth-path` filters client-side. The CLI client (`cmd/helpers/client.go`) now also promotes a JWT-shaped `WARDEN_TOKEN` (one starting with `eyJ`) to an `Authorization: Bearer` header so identity-introspection endpoints — which specifically don't read `X-Warden-Token` — work without a separate env var. Server-side fix bundled: `sys/introspect/roles` is now unauthenticated by `PathsSpecial`, matching the per-mount `auth/{jwt,cert}/introspect/roles` paths. The handler self-authenticates via the presented JWT/cert (returns 401 if neither). Without this, an agent presenting only its identity vehicle (no Warden token) got 403 from the policy check before the handler ran — defeating the whole point of introspection.
+- **`warden role list` CLI command** — agent-facing role introspection over the server's `/v1/sys/introspect/roles` aggregator. Lists the `{name, description, auth_path}` of every role the caller's identity (JWT bearer or TLS client cert) can assume across the namespace's auth mounts. Lives under a `warden role` parent group for symmetry with `warden provider list`. Composes with `--output` and `--fields`; per-mount failures surface on stderr without changing the exit code so a partial-failure mount can't mask the rest of the namespace's roles. `--auth-path` filters client-side. The CLI client (`cmd/helpers/client.go`) now also promotes a JWT-shaped `WARDEN_TOKEN` (one starting with `eyJ`) to an `Authorization: Bearer` header so identity-introspection endpoints — which specifically don't read `X-Warden-Token` — work without a separate env var. Server-side fix bundled: `sys/introspect/roles` is now unauthenticated by `PathsSpecial`, matching the per-mount `auth/{jwt,cert}/introspect/roles` paths. The handler self-authenticates via the presented JWT/cert (returns 401 if neither). Without this, an agent presenting only its identity vehicle (no Warden token) got 403 from the policy check before the handler ran — defeating the whole point of introspection.
 - **`warden schema` CLI command** — agent-facing introspection that hits the server's `/v1/sys/schema` endpoint and projects the OpenAPI doc into a friendly shape. Three modes:
   - `warden schema PATH` — describe a single path. Returns `{path, methods, description, parameters, response_schema, auth_required}` with body fields merged from `requestBody.content."application/json".schema.properties` and `sensitive` flag preserved from `x-vault-displayAttrs`. The `description` is sourced from the path-item's `description` (`HelpSynopsis`), falling back to the first operation's `summary`/`description` if the path-item has none.
   - `warden schema --list` — enumerate every path in the caller's namespace. Output is NDJSON-friendly: `{path, methods}` per line.
